Move some avro functions to AvroUtil
